Make the tree pass its own lint job, and have validate.sh run it
The CI as written could not go green. `ruff check` reported 92 errors and
`ruff format --check` wanted 15 files changed, and nothing had ever run either
command, because `scripts/validate.sh` claimed to run every check the CI runs and
did not run these two. That gap is the whole reason this went unnoticed, so it is
closed first: validate.sh now runs both, in the official ruff image, pinned to the
same version the workflow installs.

The 92 split into three kinds.

Mechanical: formatting across 15 files, plus the seven autofixable findings.
Nothing to decide, and the register table in `registers.py` was untouched, which
was the worry.

Real, and fixed:

  - `int(round(x))` in the simulator, where round already returns an int.
  - `zip(blocks, blocks[1:])` where `itertools.pairwise` says it in one word.
  - two U+2212 MINUS SIGN characters in a comment and a docstring, which look
    exactly like a hyphen and break a grep for one.
  - eight docstrings reworded to imperative mood, and nine functions that had
    none, including all seven platform `async_setup_entry` entry points.

A rule turned off with its reasoning written down, which is the only kind worth
turning off:

  - D102/D105. Fifty-eight of the 92 were Home Assistant property overrides:
    `native_value`, `is_on`, `hvac_mode`, `extra_state_attributes`. Their contract
    belongs to the base class and their subject to the class docstring, which here
    always says what the entity represents. A docstring on `native_value` inside a
    class documented as "Water delta T (outlet - inlet), design figure 5 K" can
    only repeat one of the two, and sixty of them would push the comments that do
    carry information further apart. D100, D101 and D103 stay enforced, and the
    nine D103 hits were written rather than silenced.
  - RUF012, for the same class of reason: `_attr_hvac_modes` and friends are lists
    by the base class's own annotation, so ClassVar would contradict it.

Two small things found while reading:

  - `THERMAL_POWER_UNIT` was an unused alias for `UnitOfPower.KILO_WATT`, and its
    import existed only to feed it. Both gone.
  - `SetpointDef.diagnostic` set `EntityCategory.CONFIG`, which is a different
    thing. Renamed to `config_category`.

The unused enable and command bits in `safe_write.py` stay, and now say why: the
two registers are bit fields, a bit field is only safe to write when the whole
layout is known, and that layout came out of reverse engineering a protocol that
is not published anywhere. Bit 5 of 7202 being named forced purge is why nothing
sets it by accident.

235 tests pass, and validate.sh is green on all six checks.

+++ scripts/modbus_simulator.py
@@ -214,7 +214,7 @@ class HeatPump:
             }
             # The real gateway answers unmapped addresses with zero rather than an
             # exception, and code that assumes otherwise breaks against it.
-            value = int(round(table.get(address, 0)))
+            value = round(table.get(address, 0))
             return value & 0xFFFF
 
     def write(self, address: int, value: int) -> int | None:
@@ -375,7 +375,7 @@ def main() -> int:
     # compose file does not publish the port to the host.
     parser.add_argument(
         "--host",
-        default="0.0.0.0",  # noqa: S104
+        default="0.0.0.0",
         help="address to bind (default all interfaces)",
     )
     parser.add_argument("--port", type=int, default=5020, help="TCP port (default 5020)")
